SharePointFileAnalyzer/requirements.txt
# SharePoint Integration
Office365-REST-Python-Client>=2.5.3

# AI/OpenAI Integration
openai>=1.12.0

# Excel File Generation
openpyxl>=3.1.2

# Environment Variables
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.31.0

# Document Processing
python-docx>=1.1.0
PyPDF2>=3.0.1

# Progress Indicators
tqdm>=4.66.1
